requirements.tests.txt
pytest-subprocess==1.4.2
requests-mock==1.10.0
pytest-httpserver==1.0.5